chore: change sidechannel to use promises and not async generators
This is a better api
